# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Fine Collection API
  version: 1.0.0
paths:
  /api/users/register:
    post:
      summary: Register a new user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '201':
          description: User registered
  /api/users/login:
    post:
      summary: Log in a user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Login successful, returns JWT token
  /api/items:
    get:
      summary: List all items for authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of items
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name: { type: string }
                    description: { type: string }
                    category: { type: string }
                    value: { type: number }
    post:
      summary: Create a new item
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                category: { type: string }
                value: { type: number }
      responses:
        '201':
          description: Item created
  /api/items/{id}:
    put:
      summary: Update an item
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                category: { type: string }
                value: { type: number }
      responses:
        '200':
          description: Item updated
    delete:
      summary: Delete an item
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item deleted
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT